+++ boards/wio_terminal/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Unreleased
 
+- Update `i2c_master` convenience function to use the new `sercom::v2::i2c` API
+- Add an `i2c` example
+
 # v0.5.0
 
 - Update library and examples to use `atsamd-hal` V2 APIs and upgrade BSP to Tier 1.